vect: fix type

+++ includes/vect.h
@@ -3,8 +3,8 @@
 
 typedef struct vect2
 {
-  int x;
-  int y;
+  double x;
+  double y;
 } s_vect2;
 
 
@@ -19,4 +19,4 @@ typedef s_vect2 s_vect;
 
 s_vect vect_add(s_vect a, s_vect b);
 s_vect vect_sub(s_vect a, s_vect b);
-float vect_dist(s_vect a, s_vect b);
+double vect_dist(s_vect a, s_vect b);
